// src/api/model/location.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * Property geographic location
 */
export interface Location {
  /** @minLength 1 */
  addressLine: string;
  /** @minLength 1 */
  area: string;
  /** @minLength 1 */
  district: string;
  /** @minLength 1 */
  zipCode: string;
  latitude: number;
  longitude: number;
}